+++ src/app/clusters/diagnostic-logs-server/diagnostic-logs-server.cpp
@@ -17,7 +17,7 @@
 
 #include <app/clusters/diagnostic-logs-server/diagnostic-logs-server.h>
 
-#include <app/util/af.h>
+#include <app/util/attribute-storage.h>
 #include <app/util/config.h>
 #include <lib/support/ScopedBuffer.h>
 #include <protocols/bdx/DiagnosticLogs.h>
